1. restructuring my application using blueprints into more manageable sections 2. moving configuration into its own file and created a Config class
