--- flaskapp/models.py
@@ -1,23 +1,0 @@
-from datetime import datetime
-from flaskapp import db
-
-class User(db.Model):
-    id = db.Column(db.Integer, primary_key=True)
-    username = db.Column(db.String(20), unique=True, nullable=False)
-    email = db.Column(db.String(120), unique=True, nullable=False)
-    password = db.Column(db.String(60), nullable=False)
-    notes = db.relationship('Note', backref='author', lazy=True)
-
-    def __repr__(self) -> str:
-        return f"User('{self.username}', '{self.email}')"
-    
-class Note(db.Model):
-    id = db.Column(db.Integer, primary_key=True)
-    title = db.Column(db.String(100), nullable=False)
-    date_posted = db.Column(db.DateTime, nullable=False, default=datetime.utcnow)
-    content = db.Column(db.Text, nullable=False)
-    user_id = db.Column(db.Integer, db.ForeignKey('user.id'), nullable=False)
-
-    def __repr__(self) -> str:
-        return f"Post('{self.title}', '{self.date_posted}')"
-
