--- flaskapp/routes.py
@@ -1,101 +1,0 @@
-from flask import render_template, url_for, redirect, flash, abort, request
-from flaskapp import app, db, bcrypt
-from flaskapp.forms import RegistrationForm, LoginForm, NoteForm
-from flaskapp.models import User, Note
-from flask_login import login_user, current_user, logout_user, login_required
-
-@app.route("/")
-@app.route("/home")
-def home():
-    notes = Note.query.all()
-    return render_template('home.html', notes=notes)
-
-@app.route('/about')
-def about():
-    return render_template('about.html', title='About')
-
-@app.route('/register', methods=['GET', 'POST'])
-def register():
-    if current_user.is_authenticated:
-        return redirect(url_for('home'))
-    form = RegistrationForm()
-    if form.validate_on_submit():
-        hashed_password = bcrypt.generate_password_hash(form.password.data).decode('utf-8')
-        user = User(username=form.username.data, email=form.email.data, password=hashed_password)
-        db.session.add(user)
-        db.session.commit()
-        flash('Your account was created! You are now able to log in.', 'success')
-        return redirect(url_for('login'))
-    return render_template('register.html', title='Register', form=form)
-
-@app.route('/login', methods=['GET', 'POST'])
-def login():
-    if current_user.is_authenticated:
-        return redirect(url_for('home'))
-    form = LoginForm()
-    if form.validate_on_submit():
-        user = User.query.filter_by(email=form.email.data).first()
-        if user and bcrypt.check_password_hash(user.password, form.password.data):
-            login_user(user, remember=form.remember.data)
-            flash('Login successful!', 'success')
-            return redirect(url_for('home'))
-        else:
-            flash('Login Unsuccessful. Please check email and password', 'danger')
-    return render_template('login.html', title='Login', form=form)
-
-@app.route('/logout')
-def logout():
-    logout_user()
-    return redirect(url_for('home'))
-
-@app.route('/account')
-@login_required
-def account():
-    return render_template('account.html', title='Account')
-
-@app.route('/note/new', methods=['GET', 'POST'])
-@login_required
-def new_note():
-    form = NoteForm()
-    if form.validate_on_submit():
-        note = Note(title=form.title.data, content=form.content.data, author=current_user)
-        db.session.add(note)
-        db.session.commit()
-        flash('Your note was created!', 'success')
-        return redirect(url_for('home'))
-    return render_template('create_and_update_note.html', title='New Note', form=form, legend='Create a Note')
-
-@app.route('/note/<int:note_id>')
-@login_required
-def note(note_id):
-    note = Note.query.get_or_404(note_id)
-    return render_template('note.html', title=note.title, note=note)
-
-@app.route('/note/<int:note_id>/update', methods=['GET', 'POST'])
-@login_required
-def update_note(note_id):
-    note = Note.query.get_or_404(note_id)
-    if note.author != current_user:
-        abort(403)
-    form = NoteForm()
-    if form.validate_on_submit():
-        note.title = form.title.data
-        note.content = form.content.data
-        db.session.commit()
-        flash('Your note was updated!', 'success')
-        return redirect(url_for('note', note_id=note.id))
-    elif request.method == 'GET':
-        form.title.data = note.title
-        form.content.data = note.content
-    return render_template('create_and_update_note.html', title='Update Note', form=form, legend='Update a Note')
-
-@app.route('/note/<int:note_id>/delete', methods=['POST'])
-@login_required
-def delete_note(note_id):
-    note = Note.query.get_or_404(note_id)
-    if note.author != current_user:
-        abort(403)
-    db.session.delete(note)
-    db.session.commit()
-    flash('Your note has been deleted!', 'success')
-    return redirect(url_for('home'))
