+++ flaskapp/templates/layout.html
@@ -20,25 +20,25 @@
         <header class="site-header">
             <nav class="navbar navbar-expand-md navbar-dark bg-steel fixed-top">
                 <div class="container">
-                    <a class="navbar-brand" href="{{ url_for('home') }}">AI-Powered Personal Knowledge Base</a>
+                    <a class="navbar-brand" href="{{ url_for('main.home') }}">AI-Powered Personal Knowledge Base</a>
                     <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarToggle" aria-controls="navbarToggle" aria-expanded="false" aria-label="Toggle navigation">
                         <span class="navbar-toggler-icon"></span>
                     </button>
 
                     <div class="collapse navbar-collapse" id="navbarToggle">
                         <div class="navbar-nav mr-auto">
-                            <a class="nav-item nav-link" href="{{ url_for('home') }}">Home</a>
-                            <a class="nav-item nav-link" href="{{ url_for('about') }}">About</a>
+                            <a class="nav-item nav-link" href="{{ url_for('main.home') }}">Home</a>
+                            <a class="nav-item nav-link" href="{{ url_for('main.about') }}">About</a>
                         </div>
                         <!-- Navbar Right Side -->
                         <div class="navbar-nav">
                             {% if current_user.is_authenticated %}
-                                <a class="nav-item nav-link" href="{{ url_for('new_note') }}">New Note</a>
-                                <a class="nav-item nav-link" href="{{ url_for('account') }}">Account</a>
-                                <a class="nav-item nav-link" href="{{ url_for('logout') }}">Logout</a>
+                                <a class="nav-item nav-link" href="{{ url_for('notes.new_note') }}">New Note</a>
+                                <a class="nav-item nav-link" href="{{ url_for('users.account') }}">Account</a>
+                                <a class="nav-item nav-link" href="{{ url_for('users.logout') }}">Logout</a>
                             {% else %}
-                                <a class="nav-item nav-link" href="{{ url_for('login') }}">Login</a>
-                                <a class="nav-item nav-link" href="{{ url_for('register') }}">Register</a>
+                                <a class="nav-item nav-link" href="{{ url_for('users.login') }}">Login</a>
+                                <a class="nav-item nav-link" href="{{ url_for('users.register') }}">Register</a>
                             {% endif %}
                         </div>
                     </div>
