+++ flaskapp/templates/login.html
@@ -46,7 +46,7 @@
 </div>
 <div class="border-top pt-3">
     <small class="text-muted">
-        Need An Account? <a class="ml-2" href="{{ url_for('register') }}">Sign Up Now</a>
+        Need An Account? <a class="ml-2" href="{{ url_for('users.register') }}">Sign Up Now</a>
     </small>
 </div>
 {% endblock content %}
