+++ flaskapp/templates/note.html
@@ -7,7 +7,7 @@
             <small class="text-muted">{{ note.date_posted.strftime('%Y-%m-%d') }}</small>
             {% if note.author == current_user %}
             <div>
-                <a class="btn btn-secondary btn-sm mt-1 mb-1" href="{{ url_for('update_note', note_id=note.id) }}">Update</a>
+                <a class="btn btn-secondary btn-sm mt-1 mb-1" href="{{ url_for('notes.update_note', note_id=note.id) }}">Update</a>
                 <button type="button" class="btn btn-danger btn-sm m-1" data-toggle="modal" data-target="#deleteModal">Delete</button>
             </div>
             {% endif %}
@@ -28,7 +28,7 @@
             </div>
             <div class="modal-footer">
                 <button type="button" class="btn btn-secondary" data-dismiss="modal">Close</button>
-                <form action="{{ url_for('delete_note', note_id=note.id) }}" method="POST">
+                <form action="{{ url_for('notes.delete_note', note_id=note.id) }}" method="POST">
                     <input class="btn btn-danger" type="submit" value="Delete">
                 </form>
             </div>
